+++ src/model/function.cpp
@@ -15,14 +15,14 @@ vector<string> Function::getParams (){
 	return parameters;
 }
 
-int Function::compare (ValuePtr other) {
+double Function::compare (ValuePtr other) {
 	if (getType() != other->getType())
 		return Value::compare (other);
 
 	return compare (dynamic_pointer_cast<Function>(other));
 }
 
-int Function::compare(FunctionPtr other) {
+double Function::compare(FunctionPtr other) {
 	return this - other.get();
 }
